% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CANCER_STAGING}
\alias{CANCER_STAGING}
\title{Cancer Staging}
\format{
A data frame with 1617 rows and 15 variables:
\describe{
\item{\code{value_set_name}}{character}
\item{\code{code_system}}{character}
\item{\code{logical_definition}}{character}
\item{\code{code}}{logical}
\item{\code{code_description}}{logical}
\item{\code{concept_id}}{double}
\item{\code{concept_name}}{character}
\item{\code{domain_id}}{character}
\item{\code{vocabulary_id}}{character}
\item{\code{concept_class_id}}{character}
\item{\code{standard_concept}}{character}
\item{\code{concept_code}}{character}
\item{\code{valid_start_date}}{double}
\item{\code{valid_end_date}}{double}
\item{\code{invalid_reason}}{logical}
}
}
\usage{
CANCER_STAGING
}
\description{
Minimal Common Oncology Data Elements related to AJCC Staging as
represented in the NCI Thesaurus subset of the OMOP Vocabularies.
}
\keyword{datasets}
