% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ICD10CM}
\alias{ICD10CM}
\title{icd10cm}
\format{
A data frame with 6074 rows and 15 variables:
\describe{
\item{\code{value_set_name}}{character COLUMN_DESCRIPTION}
\item{\code{code_system}}{character COLUMN_DESCRIPTION}
\item{\code{logical_definition}}{logical COLUMN_DESCRIPTION}
\item{\code{code}}{character COLUMN_DESCRIPTION}
\item{\code{code_description}}{character COLUMN_DESCRIPTION}
\item{\code{concept_id}}{double COLUMN_DESCRIPTION}
\item{\code{concept_name}}{character COLUMN_DESCRIPTION}
\item{\code{domain_id}}{character COLUMN_DESCRIPTION}
\item{\code{vocabulary_id}}{character COLUMN_DESCRIPTION}
\item{\code{concept_class_id}}{character COLUMN_DESCRIPTION}
\item{\code{standard_concept}}{logical COLUMN_DESCRIPTION}
\item{\code{concept_code}}{character COLUMN_DESCRIPTION}
\item{\code{valid_start_date}}{double COLUMN_DESCRIPTION}
\item{\code{valid_end_date}}{double COLUMN_DESCRIPTION}
\item{\code{invalid_reason}}{logical COLUMN_DESCRIPTION}
}
}
\usage{
ICD10CM
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
