% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OWL}
\alias{OWL}
\title{OWL Class Hierarchy}
\format{
An object of class \code{list} of length 21.
}
\usage{
OWL
}
\description{
All the datasets within this package are modified to reflect classes in OWL
format.
}
\details{
See the \code{mcode-rdf} vignette for more details.
}
\keyword{datasets}
