% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SNOMED}
\alias{SNOMED}
\title{SNOMED}
\format{
A data frame with 175450 rows and 18 variables:
\describe{
\item{\code{value_set_name}}{character}
\item{\code{code_system}}{character}
\item{\code{logical_definition}}{character}
\item{\code{ancestors_of}}{logical}
\item{\code{descendants_of}}{double}
\item{\code{code}}{logical}
\item{\code{code_description}}{logical}
\item{\code{all_codes}}{double}
\item{\code{concept_id}}{double}
\item{\code{concept_name}}{character}
\item{\code{domain_id}}{character}
\item{\code{vocabulary_id}}{character}
\item{\code{concept_class_id}}{character}
\item{\code{standard_concept}}{character}
\item{\code{concept_code}}{character}
\item{\code{valid_start_date}}{double}
\item{\code{valid_end_date}}{double}
\item{\code{invalid_reason}}{logical}
}
}
\usage{
SNOMED
}
\description{
All SNOMED Concepts in the Minimal Common Oncology Data Elements.
}
\keyword{datasets}
