% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPECIMEN}
\alias{SPECIMEN}
\title{Specimen}
\format{
A data frame with 3928 rows and 14 variables:
\describe{
\item{\code{concept_id}}{double}
\item{\code{value_set_name}}{character}
\item{\code{code_system}}{character}
\item{\code{logical_definition}}{logical}
\item{\code{code}}{character}
\item{\code{code_description}}{character}
\item{\code{Observation}}{character}
\item{\code{Condition}}{character}
\item{\code{Spec Anatomic Site}}{character}
\item{\code{Drug}}{character}
\item{\code{Meas Value}}{character}
\item{\code{Measurement}}{character}
\item{\code{Specimen}}{character}
\item{\code{X14}}{logical}
}
}
\usage{
SPECIMEN
}
\description{
All the possible clinical specimen types across both mCode and
OMOP.
}
\keyword{datasets}
