% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UNITS_OF_MEASUREMENT}
\alias{UNITS_OF_MEASUREMENT}
\title{Units of Measurement}
\format{
A data frame with 1011 rows and 15 variables:
\describe{
\item{\code{value_set_name}}{character}
\item{\code{code_system}}{character}
\item{\code{logical_definition}}{logical}
\item{\code{code}}{character}
\item{\code{code_description}}{character}
\item{\code{concept_id}}{double}
\item{\code{concept_name}}{character}
\item{\code{domain_id}}{character}
\item{\code{vocabulary_id}}{character}
\item{\code{concept_class_id}}{character}
\item{\code{standard_concept}}{character}
\item{\code{concept_code}}{character}
\item{\code{valid_start_date}}{double}
\item{\code{valid_end_date}}{double}
\item{\code{invalid_reason}}{character}
}
}
\usage{
UNITS_OF_MEASUREMENT
}
\description{
All the possible units of measurement across both OMOP and mCode
}
\keyword{datasets}
