% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ICD10CM}
\alias{ICD10CM}
\title{ICD-10-CM}
\format{
A data frame with 6074 rows and 15 variables:
\describe{
\item{\code{value_set_name}}{character}
\item{\code{code_system}}{character}
\item{\code{logical_definition}}{logical}
\item{\code{code}}{character}
\item{\code{code_description}}{character}
\item{\code{concept_id}}{double}
\item{\code{concept_name}}{character}
\item{\code{domain_id}}{character}
\item{\code{vocabulary_id}}{character}
\item{\code{concept_class_id}}{character}
\item{\code{standard_concept}}{logical}
\item{\code{concept_code}}{character}
\item{\code{valid_start_date}}{double}
\item{\code{valid_end_date}}{double}
\item{\code{invalid_reason}}{logical}
}
}
\usage{
ICD10CM
}
\description{
ICD-10-CM OMOP Concepts in the Minimal Common Oncology Data
Elements.
}
\keyword{datasets}
