% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MCODE_CLASS_HIERARCHY}
\alias{MCODE_CLASS_HIERARCHY}
\title{mCode-OMOP Class Hierarchy}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 215271 rows and 2 columns.
}
\usage{
MCODE_CLASS_HIERARCHY
}
\description{
All the datasets within this package are modified to reflect classes in OWL
format.
}
\details{
See the \code{mcode-rdf} vignette for more details.
}
\keyword{datasets}
