% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPECIMEN}
\alias{SPECIMEN}
\title{specimen}
\format{
A data frame with 3928 rows and 14 variables:
\describe{
\item{\code{concept_id}}{double COLUMN_DESCRIPTION}
\item{\code{value_set_name}}{character COLUMN_DESCRIPTION}
\item{\code{code_system}}{character COLUMN_DESCRIPTION}
\item{\code{logical_definition}}{logical COLUMN_DESCRIPTION}
\item{\code{code}}{character COLUMN_DESCRIPTION}
\item{\code{code_description}}{character COLUMN_DESCRIPTION}
\item{\code{Observation}}{character COLUMN_DESCRIPTION}
\item{\code{Condition}}{character COLUMN_DESCRIPTION}
\item{\code{Spec Anatomic Site}}{character COLUMN_DESCRIPTION}
\item{\code{Drug}}{character COLUMN_DESCRIPTION}
\item{\code{Meas Value}}{character COLUMN_DESCRIPTION}
\item{\code{Measurement}}{character COLUMN_DESCRIPTION}
\item{\code{Specimen}}{character COLUMN_DESCRIPTION}
\item{\code{X14}}{logical COLUMN_DESCRIPTION}
}
}
\usage{
SPECIMEN
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
